feat: add welcome message and initialize i18next in main
