+++ apps/backend/src/workers/emailWorker.ts
@@ -5,8 +5,6 @@ import { prisma } from '../lib/prisma'
 import nodemailer from 'nodemailer'
 import { appConfig } from '@/lib/appconfig'
 import i18next from 'i18next'
-import FsBackend from 'i18next-fs-backend'
-import path from 'path'
 
 const redisUrl = appConfig.REDIS_URL
 if (!redisUrl) {
@@ -27,25 +25,7 @@ const transporter = nodemailer.createTransport({
   },
 })
 
-const translationsPath = path.join(
-  __dirname,
-  __dirname.includes('dist') ? '../../..' : '../../../../',
-  'packages',
-  'shared',
-  'i18n',
-  '{{lng}}.json'
-)
 
-i18next
-  .use(FsBackend)
-  .init({
-    fallbackLng: 'en',
-    preload: ['en', 'de', 'fr', 'hu'],
-    initImmediate: false,
-    backend: {
-      loadPath: translationsPath,
-    },
-  })
 
 new Worker(
   'emails',
